Adjust objects inherit structure: now NXTexture and NXMaterial has been moved into NXObject.

+++ Nix/Core/NXGUIMaterial.cpp
@@ -365,7 +365,7 @@ void NXGUIMaterial::RenderMaterialUI_Custom_Parameters_CBufferItem(const std::st
 	{
 		// 在这里将 GUI 修改过的参数传回给材质 CBuffer，实现视觉上的变化。
 		// 实际上要拷贝的字节量是 cbDisplay 初始读取的字节数量 actualN，而不是更改 GUIStyle 以后的参数数量 N
-		UINT actualN = cbDisplay.readType;
+		UINT actualN = (UINT)cbDisplay.readType;
 		pMaterial->SetCBInfoMemoryData(cbDisplay.memoryIndex, actualN, cbDisplay.data);
 	}
 }
